py_server/Angelica_python/requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
requests
openai
flask_jwt_extended
bcrypt